Major update
basically converted the entire project to rely on javascript instead of using pyscript. py script was way too slow to generate the scrambles. Instead, i could use a javascript module instead paired with node js. this now generates the scrambles but the website relies on node js now.

Next: remove placeholder text and add a scramble as soon as the website opens instead of having to load one time first.